v1.0
Just double click it and it will do the rest (in theory).

If you need other libraries/want to add further libraries, just add their names to the list in:

set "modules=matplotlib ... sympy"

Make sure you have the correct name (check on pypi.org)

It could be that your PATH variable is not properly configured and thus pip will not be recognized as a command. Tutorial on how to fix that will be added to the readme.md.
REMIND ME OF THIS I AM FORGETFUL AS FUCK